requirements.txt
Django==1.7
django-appconf==1.0.1
django-registration-redux==1.1
pbr==0.10.7
Pillow==2.7.0
six==1.9.0
stevedore==1.2.0
virtualenv==12.0.5
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
